+++ Forge/changelog.md
@@ -1,2 +1,4 @@
-1.11.2:
-* Fix "this.scytheItem.getItem()" null pointer crash with Scythes
+1.11.3:
+* Made scythe effective against #minecraft:crops, #minecraft:flowers and #minecraft:replaceable_by_trees
+* Added Corn Seed and Cucumber Seeds to #minecraft:villager_plantable_seeds
+* Added #forge:mineable/scythe